+++ src/Smartstore.Web/wwwroot/js/smartstore.numberinput.js
@@ -63,14 +63,17 @@
             var min = $input.attr("min");
             var max = $input.attr("max");
             var step = $input.attr("step");
+            var useGrouping = $input.attr("use-grouping");
+
             min = isNaN(min) || min === "" ? -Infinity : parseFloat(min);
             max = isNaN(max) || max === "" ? Infinity : parseFloat(max);
             step = parseFloat(step) || 1;
+            useGrouping = useGrouping == 'true';
 
             var decimals = parseInt($input.data("decimals")) || 0;
             var value = parseFloat($input[0].value);
 
-            updateDisplay(value);
+            updateDisplay(value, useGrouping);
 
             this.destroy = function() {
                 $group.removeClass("numberinput-initialized");
@@ -85,7 +88,7 @@
                 var newValue = parseValue($input[0].value);
                 var focusOut = e.type === "focusout";
                 setValue(newValue, focusOut);
-                updateDisplay(newValue);
+                updateDisplay(newValue, useGrouping);
             });
 
             if (props.autoSelect) {
@@ -123,12 +126,12 @@
                     : parseInt(customFormat);
             }
 
-            function renderValue(number) {
+            function renderValue(number, useGrouping) {
                 let minDigits = Math.min(culture.numberFormat.decimals, decimals);
                 let numberFormat = new Intl.NumberFormat(culture.name, {
                     minimumFractionDigits: minDigits,
                     maximumFractionDigits: Math.max(minDigits, decimals),
-                    useGrouping: true
+                    useGrouping: useGrouping
                 });
                 return numberFormat.format(number);
             }
@@ -194,8 +197,8 @@
                 }
             }
 
-            function updateDisplay(newValue) {
-                let text = isNaN(newValue) ? "" : renderValue(newValue);
+            function updateDisplay(newValue, useGrouping) {
+                let text = isNaN(newValue) ? "" : renderValue(newValue, useGrouping);
                 $formatted.text(text);
             }