+++ changelog.md
@@ -12,6 +12,7 @@
 - Added price settings for discount requirements to be validated in product lists.
 - Faster loading of product lists that contain bundles with per-item pricing.
 - MegaSearch: a significant increase in speed, especially when dealing with large amounts of data.
+- Added MetaProperty to control number input grouping 
 
 ### Bugfixes